Added discord webhook

+++ src/main/java/de/haevn/model/DiscordWebhook.java
@@ -16,7 +16,7 @@ public class DiscordWebhook {
         private String url;
         private String title;
         private String description;
-        private int color;
+        private long color;
     }
 
 }
